Scheduling and Emailing
General -
Did a bit of cleaning of unused imports and changed some typos in
variables' names.

All DB types -
Added a (String path) constructor to the father class and changed the
path to the one from the config file for all

RemindersDB -
Changed the reminding_time to date type for convenience.

Reminder -
Added a Date type reminding date for convenience due to the change in
the table.

webServer -
Passing the ConfigManager to the SocketThreads, using it to create the
DB's.
Added mails to pass due dated reminders but it will probably not be used
in the end.

Consts -
Added the new config key's and some consts for the mails.

HttpResponse -
Added the send mails logic to Tasks and Polls  creators.
Changed update poll logic a bit.

HttpRequest -
Added a constructor which gets ConfigManager instead of defaultpage and
root.
Added set timers for reminders and tasks methods.
Added send mails to poll participants.

SMTPClients -
Fixed some bugs.
Currently prints all the process, not really necessary.

TODO:
- Change all the usage of SMTPClient to work with the Config settings
and not hard coded
- Check and fix the problems with ID's
- Change the linking inside the Task mail, it's currently not right.
- There is a problem with "MailPollParticipants" at the moment. Need to
fix it and check if the poll links work.
- The Task status currently does not change to "over due time" when the
time is due. Probably connected to the Id problem.