Bug fixing in Emailing
General -
No major changes in general structure. Fixed specific bugs and changed
some logic in mailing system.

HttpResponse -
Added a "answerer" param to poll_reply.html to know who replied to the
poll
Changed the content of the mail sent to poll creator when the poll is
answered

HttpRequest -
Fixed the mail to task recipient content
Fixed the problem with "-1" Id's
Changed the logic of the creation of the content of the mail sent to
poll recipients due to change in paramaters

PollDB and TaskDB -
When creating a row return it's newly created Id because the object that
it is created by does not have it

SMTPClient -
Changed "Mr. Tasker" to "Mr.Tasker" due to (unknown sender) problem.